Final Push From Me
This is the final push from me for a while, everything now works but it isn't pretty. Flights' table still needs some more flights but that's easy. Just use the register function to slam them in. It doesn't look pretty, but it should pass all test cases and requirements for this project. It isn't very safe though.